Add basic Consul support to calico-docker
Add functionality to create and remove services in Consul. Services
are added with the "-direct" suffix in order to differentiate from other
services. If you do a DNS lookup against the "-direct" service, IPs
from the Calico network will be returned. Thus, we obtain an option to
query services and discover their IPs from the Calico network.
Services are added and removed using the Powerstrip post-hook rules.